// types.ts
export interface TrailerModel {
  id: string;
  name: string;
  imageUrl: string;
}

export interface ColorTheme {
  id: string;
  name: string;
  description: string;
  gradientText: string;
  gradientBorder: string;
  glow: string;
}

export interface BrandingDetails {
  slogan: string;
  website: string;
  phone: string;
}